refactor: moved rendering stars in review form into separate component
