+++ src/components/review-form/review-form.tsx
@@ -1,14 +1,7 @@
-import { ChangeEvent, Fragment, ReactEventHandler, useState } from 'react';
+import { ChangeEvent, useState } from 'react';
 import { MAX_COMMENT_LENGTH, MIN_COMMENT_LENGTH } from '../../const';
-import { RatingNumber, RatingString } from '../../types/rating';
-
-const ratingMap: Array<{ title: RatingString; value: RatingNumber }> = [
-  { title: 'excellent', value: 5 },
-  { title: 'good', value: 4 },
-  { title: 'not bad', value: 3 },
-  { title: 'bad', value: 2 },
-  { title: 'terrible', value: 1 },
-];
+import { RatingNumber } from '../../types/rating';
+import ReviewFormRating from '../review-form-rating/review-form-rating';
 
 function ReviewForm() {
   const [comment, setComment] = useState<{
@@ -21,11 +14,12 @@ function ReviewForm() {
     comment.review.length <= MAX_COMMENT_LENGTH &&
     comment.rating !== 0;
 
-  const handleChange: ReactEventHandler<
-    HTMLInputElement | HTMLTextAreaElement
-  > = (evt: ChangeEvent<HTMLTextAreaElement>) => {
-    const { name, value } = evt.currentTarget;
-    setComment({ ...comment, [name]: value });
+  const handleRatingChange = (value: RatingNumber) => {
+    setComment({ ...comment, rating: value });
+  };
+
+  const handleTextareaChange = (evt: ChangeEvent<HTMLTextAreaElement>) => {
+    setComment({ ...comment, review: evt.currentTarget.value });
   };
 
   return (
@@ -33,36 +27,14 @@ function ReviewForm() {
       <label className="reviews__label form__label" htmlFor="review">
         Your review
       </label>
-      <div className="reviews__rating-form form__rating">
-        {ratingMap.map(({ title, value }) => (
-          <Fragment key={value}>
-            <input
-              className="form__rating-input visually-hidden"
-              name="rating"
-              defaultValue={value}
-              id={`${value}-stars`}
-              type="radio"
-              onChange={handleChange}
-            />
-            <label
-              htmlFor={`${value}-stars`}
-              className="reviews__rating-label form__rating-label"
-              title={title}
-            >
-              <svg className="form__star-image" width="37" height="33">
-                <use xlinkHref="#icon-star"></use>
-              </svg>
-            </label>
-          </Fragment>
-        ))}
-      </div>
+      <ReviewFormRating handleChange={handleRatingChange} />
       <textarea
         className="reviews__textarea form__textarea"
         id="review"
         name="review"
         placeholder="Tell how was your stay, what you like and what can be improved"
         defaultValue={comment.review}
-        onChange={handleChange}
+        onChange={handleTextareaChange}
       />
       <div className="reviews__button-wrapper">
         <p className="reviews__help">
